fix(editor core): click outside handler
